+++ lib/ui/academics/SkillsSelfEvaluationInfoPanel.dart
@@ -146,7 +146,7 @@ class SkillsSelfEvaluationInfoPanel extends StatelessWidget {
   }
 
   void _onTapLink(BuildContext context, SkillsSelfEvaluationLink link) {
-    if (link.internal && UrlUtils.launchInternal(link.url)) {
+    if (link.internal && UrlUtils.canLaunchInternal(link.url)) {
       Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: link.url)));
     } else if (link.url != null) {
       Uri? parsedUri = Uri.tryParse(link.url!);