+++ lib/ui/canvas/CanvasSyllabusHtmlPanel.dart
@@ -98,7 +98,7 @@ class _CanvasSyllabusHtmlPanelState extends State<CanvasSyllabusHtmlPanel> {
 
   void _onTapLink(String? url) {
     if (StringUtils.isNotEmpty(url)) {
-      if (UrlUtils.launchInternal(url)) {
+      if (UrlUtils.canLaunchInternal(url)) {
         Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: url, analyticsFeature: widget.analyticsFeature,)));
       } else {
         Uri? uri = Uri.tryParse(url!);