+++ lib/ui/home/HomeSafeWalkRequestWidget.dart
@@ -72,7 +72,7 @@ class HomeSafeWalkRequestWidget extends StatelessWidget {
       if (DeepLink().isAppUrl(url)) {
         DeepLink().launchUrl(url);
       }
-      else if (launchInternal && UrlUtils.launchInternal(url)){
+      else if (launchInternal && UrlUtils.canLaunchInternal(url)){
         Navigator.push(context, CupertinoPageRoute(builder: (context) => WebPanel(url: url)));
       }
       else {